+++ lib/src/environment.dart
@@ -5,7 +5,7 @@
 // DO NOT EDIT. This file was generated from async_environment.dart.
 // See tool/grind/synchronize.dart for details.
 //
-// Checksum: 2522d5fbfb9d301b361a9bacac97228de2c6fd68
+// Checksum: 0459cbe5c439f3d45f24b739ed1f36b517d338b8
 //
 // ignore_for_file: unused_import
 
@@ -23,6 +23,7 @@ import 'exception.dart';
 import 'extend/extender.dart';
 import 'module.dart';
 import 'module/forwarded_view.dart';
+import 'module/shadowed_view.dart';
 import 'util/merged_map_view.dart';
 import 'util/public_member_map_view.dart';
 import 'utils.dart';
@@ -47,6 +48,13 @@ class Environment {
   ///
   /// This is `null` if there are no forwarded modules.
   List<Module<Callable>> _forwardedModules;
+
+  /// Modules forwarded by nested imports at each lexical scope level *beneath
+  /// the global scope*.
+  ///
+  /// This is `null` until it's needed, since most environments won't ever use
+  /// this.
+  List<List<Module<Callable>>> _nestedForwardedModules;
 
   /// Modules from [_modules], [_globalModules], and [_forwardedModules], in the
   /// order in which they were `@use`d.
@@ -134,6 +142,7 @@ class Environment {
       : _modules = {},
         _globalModules = null,
         _forwardedModules = null,
+        _nestedForwardedModules = null,
         _allModules = [],
         _variables = [{}],
         _variableNodes = sourceMap ? [{}] : null,
@@ -147,6 +156,7 @@ class Environment {
       this._modules,
       this._globalModules,
       this._forwardedModules,
+      this._nestedForwardedModules,
       this._allModules,
       this._variables,
       this._variableNodes,
@@ -170,6 +180,7 @@ class Environment {
       _modules,
       _globalModules,
       _forwardedModules,
+      _nestedForwardedModules,
       _allModules,
       _variables.toList(),
       _variableNodes?.toList(),
@@ -183,6 +194,7 @@ class Environment {
   /// functions, and mixins, but not its modules.
   Environment global() => Environment._(
       {},
+      null,
       null,
       null,
       [],
@@ -276,30 +288,63 @@ class Environment {
   /// This is called when [module] is `@import`ed.
   void importForwards(Module<Callable> module) {
     if (module is _EnvironmentModule) {
-      for (var forwarded in module._environment._forwardedModules ??
-          const <Module<Callable>>[]) {
-        _globalModules ??= {};
-        _globalModules.add(forwarded);
+      var forwarded = module._environment._forwardedModules;
+      if (forwarded == null) return;
 
-        // Remove existing definitions that the forwarded members are now
-        // shadowing.
-        for (var variable in forwarded.variables.keys) {
-          var index =
-              _variableIndices.remove(variable) ?? _variableIndex(variable);
-          if (index != null) {
-            _variables[index].remove(variable);
-            if (_variableNodes != null) _variableNodes[index].remove(variable);
+      _globalModules ??= {};
+      _forwardedModules ??= [];
+
+      var forwardedVariableNames =
+          forwarded.expand((module) => module.variables.keys).toSet();
+      var forwardedFunctionNames =
+          forwarded.expand((module) => module.functions.keys).toSet();
+      var forwardedMixinNames =
+          forwarded.expand((module) => module.mixins.keys).toSet();
+
+      if (atRoot) {
+        // Hide members from modules that have already been imported or
+        // forwarded that would otherwise conflict with the @imported members.
+        for (var module in _globalModules.toList()) {
+          var shadowed = ShadowedModuleView.ifNecessary(module,
+              variables: forwardedVariableNames,
+              mixins: forwardedMixinNames,
+              functions: forwardedFunctionNames);
+          if (shadowed != null) {
+            _globalModules.remove(module);
+            _globalModules.add(shadowed);
           }
         }
-        for (var function in forwarded.functions.keys) {
-          var index =
-              _functionIndices.remove(function) ?? _functionIndex(function);
-          if (index != null) _functions[index].remove(function);
+        for (var i = 0; i < _forwardedModules.length; i++) {
+          var module = _forwardedModules[i];
+          var shadowed = ShadowedModuleView.ifNecessary(module,
+              variables: forwardedVariableNames,
+              mixins: forwardedMixinNames,
+              functions: forwardedFunctionNames);
+          if (shadowed != null) _forwardedModules[i] = shadowed;
         }
-        for (var mixin in forwarded.mixins.keys) {
-          var index = _mixinIndices.remove(mixin) ?? _mixinIndex(mixin);
-          if (index != null) _mixins[index].remove(mixin);
-        }
+
+        _globalModules.addAll(forwarded);
+        _forwardedModules.addAll(forwarded);
+      } else {
+        _nestedForwardedModules ??=
+            List.generate(_variables.length - 1, (_) => []);
+        _nestedForwardedModules.last.addAll(forwarded);
+      }
+
+      // Remove existing member definitions that are now shadowed by the
+      // forwarded modules.
+      for (var variable in forwardedVariableNames) {
+        _variableIndices.remove(variable);
+        _variables.last.remove(variable);
+        if (_variableNodes != null) _variableNodes.last.remove(variable);
+      }
+      for (var function in forwardedFunctionNames) {
+        _functionIndices.remove(function);
+        _functions.last.remove(function);
+      }
+      for (var mixin in forwardedMixinNames) {
+        _mixinIndices.remove(mixin);
+        _mixins.last.remove(mixin);
       }
     }
   }
@@ -471,6 +516,19 @@ class Environment {
       _variables.first[name] = value;
       if (_variableNodes != null) _variableNodes.first[name] = nodeWithSpan;
       return;
+    }
+
+    if (_nestedForwardedModules != null &&
+        !_variableIndices.containsKey(name) &&
+        _variableIndex(name) == null) {
+      for (var modules in _nestedForwardedModules.reversed) {
+        for (var module in modules.reversed) {
+          if (module.variables.containsKey(name)) {
+            module.setVariable(name, value, nodeWithSpan);
+            return;
+          }
+        }
+      }
     }
 
     var index = _lastVariableName == name
@@ -656,6 +714,7 @@ class Environment {
     _variableNodes?.add({});
     _functions.add({});
     _mixins.add({});
+    _nestedForwardedModules?.add([]);
     try {
       return callback();
     } finally {
@@ -671,14 +730,18 @@ class Environment {
       for (var name in _mixins.removeLast().keys) {
         _mixinIndices.remove(name);
       }
+      _nestedForwardedModules?.removeLast();
     }
   }
 
   /// Returns a module that represents the top-level members defined in [this],
   /// that contains [css] as its CSS tree, which can be extended using
   /// [extender].
-  Module<Callable> toModule(CssStylesheet css, Extender extender) =>
-      _EnvironmentModule(this, css, extender, forwarded: _forwardedModules);
+  Module<Callable> toModule(CssStylesheet css, Extender extender) {
+    assert(atRoot);
+    return _EnvironmentModule(this, css, extender,
+        forwarded: _forwardedModules);
+  }
 
   /// Returns the module with the given [namespace], or throws a
   /// [SassScriptException] if none exists.
@@ -691,7 +754,8 @@ class Environment {
   }
 
   /// Returns the result of [callback] if it returns non-`null` for exactly one
-  /// module in [_globalModules].
+  /// module in [_globalModules] *or* for any module in
+  /// [_nestedForwardedModules].
   ///
   /// Returns `null` if [callback] returns `null` for all modules. Throws an
   /// error if [callback] returns non-`null` for more than one module.
@@ -699,6 +763,15 @@ class Environment {
   /// The [type] should be the singular name of the value type being returned.
   /// It's used to format an appropriate error message.
   T _fromOneModule<T>(String type, T callback(Module<Callable> module)) {
+    if (_nestedForwardedModules != null) {
+      for (var modules in _nestedForwardedModules.reversed) {
+        for (var module in modules.reversed) {
+          var value = callback(module);
+          if (value != null) return value;
+        }
+      }
+    }
+
     if (_globalModules == null) return null;
 
     T value;