+++ lib/src/environment.dart
@@ -5,7 +5,7 @@
 // DO NOT EDIT. This file was generated from async_environment.dart.
 // See tool/grind/synchronize.dart for details.
 //
-// Checksum: 0459cbe5c439f3d45f24b739ed1f36b517d338b8
+// Checksum: b52f5d921ea089ce2924b43ada11b5c61f15f3fa
 //
 // ignore_for_file: unused_import
 
@@ -19,6 +19,7 @@ import 'ast/css.dart';
 import 'ast/node.dart';
 import 'ast/sass.dart';
 import 'callable.dart';
+import 'configuration.dart';
 import 'exception.dart';
 import 'extend/extender.dart';
 import 'module.dart';
@@ -732,6 +733,22 @@ class Environment {
       }
       _nestedForwardedModules?.removeLast();
     }
+  }
+
+  /// Creates an implicit configuration from the variables declared in this
+  /// environment.
+  Configuration toImplicitConfiguration() {
+    var configuration = <String, ConfiguredValue>{};
+    for (var i = 0; i < _variables.length; i++) {
+      var values = _variables[i];
+      var nodes =
+          _variableNodes == null ? <String, AstNode>{} : _variableNodes[i];
+      for (var name in values.keys) {
+        configuration[name] =
+            ConfiguredValue(values[name], nodes[name]?.span, nodes[name]);
+      }
+    }
+    return Configuration(configuration, isImplicit: true);
   }
 
   /// Returns a module that represents the top-level members defined in [this],