+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 * **Potentially breaking bug fix:** Members loaded through a nested `@import`
   are no longer ever accessible outside that nested context.
 
+* Don't throw an error when importing two modules that both forward members with
+  the same name. The latter name now takes precedence over the former, as per
+  the specification.
+
 ## 1.23.7
 
 * No user-visible changes.