+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.24.0
+
+* Support configuring modules through `@import` rules.
+
 ## 1.23.8
 
 * **Potentially breaking bug fix:** Members loaded through a nested `@import`
@@ -9,7 +13,7 @@
 
 ## 1.23.7
 
-* No user-visible changes.
+* No user-visible changes
 
 ## 1.23.6